refactor: change nest preprocessing location
