+++ now/apps/text_to_text_and_image/app.py
@@ -2,8 +2,7 @@ import os
 from typing import Dict, Optional
 
 from docarray import DocumentArray
-from now_common.preprocess import preprocess_text
-from now_common.utils import preprocess_nested_docs
+from now_common.preprocess import preprocess_nested_docs, preprocess_text
 
 from now.apps.base.app import JinaNOWApp
 from now.constants import Apps, Modalities
