+++ now/constants.py
@@ -6,7 +6,7 @@ from now.utils import BetterEnum
 DEMO_DATASET_DOCARRAY_VERSION = '0.13.17'
 
 DOCKER_BFF_PLAYGROUND_TAG = '0.0.117-refactor-bump-jina-version'
-NOW_PREPROCESSOR_VERSION = '0.0.73-devnest1'
+NOW_PREPROCESSOR_VERSION = '0.0.73-devnest2'
 NOW_AUTH_EXECUTOR_VERSION = '0.0.3-8'
