+++ now_common/preprocess.py
@@ -1,6 +1,8 @@
 from docarray import Document, DocumentArray
 
 from now.data_loading.convert_datasets_to_jpeg import to_thumbnail_jpg
+from now.data_loading.utils import transform_es_data
+from now.now_dataclasses import UserInput
 
 
 def preprocess_images(da: DocumentArray) -> DocumentArray:
@@ -67,3 +69,27 @@ def preprocess_text(da: DocumentArray, split_by_sentences=False) -> DocumentArra
         da = DocumentArray(d for d in gen_split_by_sentences())
 
     return DocumentArray(d for d in da if d.text and d.text != '')
+
+
+def preprocess_nested_docs(da: DocumentArray, user_input: UserInput) -> DocumentArray:
+    """
+    Process a `DocumentArray` with `Document`s that have `chunks` of nested `Document`s.
+    It constructs `Document`s containg two chunks: one containing image data and another
+    containing text data. Fields for indexing should be specified in the `UserInput`.
+
+    :param da: A `DocumentArray` containing nested chunks.
+    :return: A `DocumentArray` with `Document`s containing text and image chunks.
+    """
+    fields = user_input.task.indexer_scope
+    transformed_da = transform_es_data(da)
+    return DocumentArray(
+        [
+            Document(
+                chunks=[
+                    Document(text=doc[fields['text']][0]),
+                    Document(uri=doc[fields['image']][0]),
+                ]
+            )
+            for doc in transformed_da
+        ]
+    )
