+++ now_common/utils.py
@@ -4,11 +4,10 @@ from os.path import expanduser as user
 from typing import Dict, Optional
 
 import hubble
-from docarray import Document, DocumentArray
+from docarray import DocumentArray
 
 from now.apps.base.app import JinaNOWApp
 from now.constants import NOW_PREPROCESSOR_VERSION, PREFETCH_NR
-from now.data_loading.utils import transform_es_data
 from now.finetuning.run_finetuning import finetune
 from now.finetuning.settings import FinetuneSettings, parse_finetune_settings
 from now.now_dataclasses import UserInput
@@ -121,30 +120,6 @@ def common_setup(
     )
 
     return env_dict
-
-
-def preprocess_nested_docs(da: DocumentArray, user_input: UserInput) -> DocumentArray:
-    """
-    Process a `DocumentArray` with `Document`s that have `chunks` of nested `Document`s.
-    It constructs `Document`s containg two chunks: one containing image data and another
-    containing text data. Fields for indexing should be specified in the `UserInput`.
-
-    :param da: A `DocumentArray` containing nested chunks.
-    :return: A `DocumentArray` with `Document`s containing text and image chunks.
-    """
-    fields = user_input.task.indexer_scope
-    transformed_da = transform_es_data(da)
-    return DocumentArray(
-        [
-            Document(
-                chunks=[
-                    Document(text=doc[fields['text']][0]),
-                    Document(uri=doc[fields['image']][0]),
-                ]
-            )
-            for doc in transformed_da
-        ]
-    )
 
 
 def _get_email():
